+++ dev/payload-types.ts
@@ -130,6 +130,21 @@ export interface UserAuthOperations {
 export interface Post {
   id: number;
   title?: string | null;
+  body?: {
+    root: {
+      type: string;
+      children: {
+        type: any;
+        version: number;
+        [k: string]: unknown;
+      }[];
+      direction: ('ltr' | 'rtl') | null;
+      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
+      indent: number;
+      version: number;
+    };
+    [k: string]: unknown;
+  } | null;
   layout?:
     | (
         | {
@@ -278,6 +293,7 @@ export interface PayloadMigration {
  */
 export interface PostsSelect<T extends boolean = true> {
   title?: T;
+  body?: T;
   layout?:
     | T
     | {